Extract class to single file.
