+++ src/MahApps.Metro/MahApps.Metro.Shared/Converters/ThicknessBindingConverter.cs
@@ -45,67 +45,4 @@ namespace MahApps.Metro.Converters
             return DependencyProperty.UnsetValue;
         }
     }
-
-    [ValueConversion(typeof(Thickness), typeof(double), ParameterType = typeof(ThicknessSideType))]
-    public class BorderThicknessToStrokeThicknessConverter : IValueConverter
-    {
-        public ThicknessSideType TakeThicknessSide { get; set; }
-
-        public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
-        {
-            if (value is Thickness)
-            {
-                // yes, we can override it with the parameter value
-                if (parameter is ThicknessSideType)
-                {
-                    this.TakeThicknessSide = (ThicknessSideType)parameter;
-                }
-                var orgThickness = (Thickness)value;
-                switch (this.TakeThicknessSide)
-                {
-                    case ThicknessSideType.Left:
-                        return orgThickness.Left;
-                    case ThicknessSideType.Top:
-                        return orgThickness.Top;
-                    case ThicknessSideType.Right:
-                        return orgThickness.Right;
-                    case ThicknessSideType.Bottom:
-                        return orgThickness.Bottom;
-                    default:
-                        return default(double);
-                }
-            }
-
-            return default(double);
-        }
-
-        public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
-        {
-            throw new InvalidOperationException("Cannot convert back to source");
-        }
-    }
-
-    public enum ThicknessSideType
-    {
-        /// <summary>
-        /// Use all sides.
-        /// </summary>
-        None,
-        /// <summary>
-        /// Ignore the left side.
-        /// </summary>
-        Left,
-        /// <summary>
-        /// Ignore the top side.
-        /// </summary>
-        Top,
-        /// <summary>
-        /// Ignore the right side.
-        /// </summary>
-        Right,
-        /// <summary>
-        /// Ignore the bottom side.
-        /// </summary>
-        Bottom
-    }
 }
