+++ src/MahApps.Metro/MahApps.Metro.Shared/Converters/ThicknessSideType.cs
@@ -1,0 +1,26 @@
+namespace MahApps.Metro.Converters
+{
+    public enum ThicknessSideType
+    {
+        /// <summary>
+        /// Use all sides.
+        /// </summary>
+        None,
+        /// <summary>
+        /// Ignore the left side.
+        /// </summary>
+        Left,
+        /// <summary>
+        /// Ignore the top side.
+        /// </summary>
+        Top,
+        /// <summary>
+        /// Ignore the right side.
+        /// </summary>
+        Right,
+        /// <summary>
+        /// Ignore the bottom side.
+        /// </summary>
+        Bottom
+    }
+}
